Extract Settlement table model into a separate class
This is to make the core config editor smaller in prepration of arriving
settlement logic.

#431
