+++ mars-sim-ui/src/main/java/org/mars_sim/msp/ui/swing/configeditor/SimulationConfigEditor.java
@@ -20,11 +20,7 @@ import java.awt.event.ItemListener;
 import java.awt.event.WindowAdapter;
 import java.awt.event.WindowEvent;
 import java.util.ArrayList;
-import java.util.Collections;
-import java.util.HashSet;
-import java.util.Iterator;
 import java.util.List;
-import java.util.Set;
 import java.util.logging.Level;
 
 import javax.swing.BorderFactory;
@@ -43,7 +39,8 @@ import javax.swing.SwingUtilities;
 import javax.swing.WindowConstants;
 import javax.swing.border.EmptyBorder;
 import javax.swing.event.ListDataListener;
-import javax.swing.table.AbstractTableModel;
+import javax.swing.event.TableModelEvent;
+import javax.swing.event.TableModelListener;
 import javax.swing.table.DefaultTableCellRenderer;
 import javax.swing.table.TableCellEditor;
 import javax.swing.table.TableColumn;
@@ -62,7 +59,6 @@ import org.mars_sim.msp.core.logging.SimLogger;
 import org.mars_sim.msp.core.person.Crew;
 import org.mars_sim.msp.core.person.CrewConfig;
 import org.mars_sim.msp.core.person.PersonConfig;
-import org.mars_sim.msp.core.reportingAuthority.ReportingAuthority;
 import org.mars_sim.msp.core.reportingAuthority.ReportingAuthorityFactory;
 import org.mars_sim.msp.core.structure.InitialSettlement;
 import org.mars_sim.msp.core.structure.SettlementConfig;
@@ -146,440 +142,12 @@ public class SimulationConfigEditor {
 		}
 	}
 	
-	/**
-	 * Inner class for the settlement table model.
-	 */
-	@SuppressWarnings("serial")
-	private class SettlementTableModel extends AbstractTableModel {
 
-		// Inner class representing a settlement configuration.
-		private final class SettlementInfo {
-			String name;
-			String sponsor;
-			String template;
-			String population;
-			String numOfRobots;
-			String latitude;
-			String longitude;
-			String crew;
-		}
-
-		private String[] columns;
-		private List<SettlementInfo> settlementInfoList;
-		
-		/**
-		 * Hidden Constructor.
-		 */
-		private SettlementTableModel() {
-			super();
-
-			// Add table columns.
-			columns = new String[] { 
-					Msg.getString("SimulationConfigEditor.column.name"), //$NON-NLS-1$
-					Msg.getString("SimulationConfigEditor.column.sponsor"), //$NON-NLS-1$
-					Msg.getString("SimulationConfigEditor.column.template"), //$NON-NLS-1$
-					Msg.getString("SimulationConfigEditor.column.population"), //$NON-NLS-1$
-					Msg.getString("SimulationConfigEditor.column.crew"),
-					Msg.getString("SimulationConfigEditor.column.numOfRobots"), //$NON-NLS-1$
-					Msg.getString("SimulationConfigEditor.column.latitude"), //$NON-NLS-1$
-					Msg.getString("SimulationConfigEditor.column.longitude") //$NON-NLS-1$
-			};
-			
-			this.settlementInfoList = new ArrayList<>();
-		}
-
-		
-		/**
-		 * Load the default settlements in the table.
-		 */
-		public void loadDefaultSettlements(Scenario selected) {
-			settlementInfoList.clear();
-			boolean hasSponsor = false;
-			String sponsorCC = null;
-			List<String> usedNames = new ArrayList<>();
-			
-			if (mode == GameMode.COMMAND) {
-				sponsorCC = personConfig.getCommander().getSponsorStr();
-				logger.config("The commander's sponsor is " + sponsorCC + ".");
-			}
-			
-			for (InitialSettlement spec : selected.getSettlements()) {
-				SettlementInfo info = toSettlementInfo(spec);
-
-				// Save this name to the list
-				usedNames.add(info.name);
-							
-				// Modify the sponsor in case of the Commander Mode
-				if (mode == GameMode.COMMAND) {
-					if (sponsorCC.equals(info.sponsor)) {
-						hasSponsor = true;
-					}
-				}
-					
-				settlementInfoList.add(info);
-			}
-			
-			if (mode == GameMode.COMMAND) {
-				
-				if (!hasSponsor) {
-					// Change the 1st settlement's sponsor to match that of the commander
-					settlementInfoList.get(0).sponsor = sponsorCC;
-					
-					// Gets a list of settlement names that are tailored to this country
-					ReportingAuthority ra = raFactory.getItem(sponsorCC);
-					List<String> candidateNames = new ArrayList<>(ra.getSettlementNames());
-					Collections.shuffle(candidateNames);
-					candidateNames.removeAll(usedNames);
-					if (!candidateNames.isEmpty()) {
-						settlementInfoList.get(0).name = candidateNames.get(0);
-					}
-					logger.config( 
-							"The 1st settlement's sponsor be matched to the commander's sponsor.");
-				}
-				
-				else {
-					logger.config( 
-							"The commander's sponsor will match one of the settlements' sponsor in the site editor.");
-				}
-			}
-				
-			fireTableDataChanged();
-		}
-
-		private SettlementInfo toSettlementInfo(InitialSettlement spec) {
-			SettlementInfo info = new SettlementInfo();
-			info.name = spec.getName();
-			info.sponsor = spec.getSponsor();
-			info.crew = spec.getCrew();
-			info.template = spec.getSettlementTemplate();
-			info.population = Integer.toString(spec.getPopulationNumber());
-			info.numOfRobots = Integer.toString(spec.getNumOfRobots());
-			Coordinates location = spec.getLocation();
-			if (location != null) {
-				info.latitude = location.getFormattedLatitudeString();
-				info.longitude = location.getFormattedLongitudeString();
-			}	
-			return info;
-		}
-
-
-		/**
-		 * Get the rows as InitialSettlements.
-		 * @return
-		 */
-		public List<InitialSettlement> getSettlements() {
-
-			List<InitialSettlement> is = new ArrayList<>();
-
-			// Add configuration settlements from table data.
-			for (SettlementInfo info : settlementInfoList) {
-				int populationNum = Integer.parseInt(info.population);
-				int numOfRobots = Integer.parseInt(info.numOfRobots);
-				
-				// take care to internationalize the coordinates
-				String latitude = info.latitude.replace("N", Msg.getString("direction.northShort")); //$NON-NLS-1$ //$NON-NLS-2$
-				latitude = latitude.replace("S", Msg.getString("direction.southShort")); //$NON-NLS-1$ //$NON-NLS-2$
-				String longitude = info.longitude.replace("E", Msg.getString("direction.eastShort")); //$NON-NLS-1$ //$NON-NLS-2$
-				longitude = longitude.replace("W", Msg.getString("direction.westShort")); //$NON-NLS-1$ //$NON-NLS-2$
-
-				Coordinates location = new Coordinates(latitude, longitude);
-
-				is.add(new InitialSettlement(info.name, info.sponsor, info.template,
-													populationNum, numOfRobots, location, info.crew));
-
-			}
-			
-			return is;
-		}
-		
-		@Override
-		public int getRowCount() {
-			return settlementInfoList.size();
-		}
-
-		@Override
-		public int getColumnCount() {
-			return columns.length;
-		}
-
-		/*
-		 * JTable uses this method to determine the default renderer/ editor for each
-		 * cell. If we didn't implement this method, then the last column would contain
-		 * text ("true"/"false"), rather than a check box.
-		 */
-		public Class<?> getColumnClass(int c) {
-			return getValueAt(0, c).getClass();
-		}
-
-		@Override
-		public String getColumnName(int columnIndex) {
-			if ((columnIndex > -1) && (columnIndex < columns.length)) {
-				return columns[columnIndex];
-			} else {
-				return Msg.getString("SimulationConfigEditor.log.invalidColumn"); //$NON-NLS-1$
-			}
-		}
-
-		@Override
-		public boolean isCellEditable(int row, int column) {
-			return true;
-		}
-
-		@Override
-		public Object getValueAt(int row, int column) {
-			Object result = Msg.getString("unknown"); //$NON-NLS-1$
-
-			if ((row > -1) && (row < getRowCount())) {
-				SettlementInfo info = settlementInfoList.get(row);
-				if ((column > -1) && (column < getColumnCount())) {
-					switch (column) {
-					case SETTLEMENT_COL:
-						result = info.name;
-						break;
-					case SPONSOR_COL:
-						result = info.sponsor;
-						break;
-					case PHASE_COL:
-						result = info.template;
-						break;
-					case SETTLER_COL:
-						result = info.population;
-						break;
-					case CREW_COL:
-						result = info.crew;
-						break;
-					case BOT_COL:
-						result = info.numOfRobots;
-						break;
-					case LAT_COL:
-						result = info.latitude;
-						break;
-					case LON_COL:
-						result = info.longitude;
-						break;
-					}
-				} else {
-					result = Msg.getString("SimulationConfigEditor.log.invalidColumn"); //$NON-NLS-1$
-				}
-			} else {
-				result = Msg.getString("SimulationConfigEditor.log.invalidRow"); //$NON-NLS-1$
-			}
-
-			return result;
-		}
-
-		@Override
-		public void setValueAt(Object aValue, int rowIndex, int columnIndex) {
-			if ((rowIndex > -1) && (rowIndex < getRowCount())) {
-				SettlementInfo info = settlementInfoList.get(rowIndex);
-				if ((columnIndex > -1) && (columnIndex < getColumnCount())) {
-					switch (columnIndex) {
-					
-					case SETTLEMENT_COL:
-						info.name = (String) aValue;
-						break;
-						
-					case SPONSOR_COL:
-						String newSponsor = (String) aValue;
-						if (!info.sponsor.equals(newSponsor)) {
-							info.sponsor = newSponsor;
-							String newName = tailorSettlementNameBySponsor(info.sponsor, rowIndex);
-							if (newName != null) {
-								info.name = newName;
-							}
-						}
-						break;	
-						
-					case PHASE_COL:
-						info.template = (String) aValue;
-						info.population = Integer.toString(determineNewSettlementPopulation(info.template));
-						info.numOfRobots = Integer.toString(determineNewSettlementNumOfRobots(info.template));
-						break;
-						
-					case SETTLER_COL:
-						info.population = (String) aValue;
-						break;
-						
-					case CREW_COL:
-						info.crew = (String) aValue;
-						break;
-						
-					case BOT_COL:
-						info.numOfRobots = (String) aValue;
-						break;
-
-					case LAT_COL:
-						String latStr = ((String) aValue).trim();
-						double doubleLat = 0;
-						String dir1 = latStr.substring(latStr.length() - 1, latStr.length());
-						if (dir1.equalsIgnoreCase("N") || dir1.equalsIgnoreCase("S")) {
-							if (latStr.length() > 2) {
-								doubleLat = Double.parseDouble(latStr.substring(0, latStr.length() - 1));
-								doubleLat = Math.round(doubleLat * 100.0) / 100.0;
-								info.latitude = doubleLat + " " + dir1.toUpperCase();
-							}
-							else {
-								info.latitude = (String) aValue;
-							}
-						}
-						else {
-							info.latitude = (String) aValue;
-						}
-						String latError = Coordinates.checkLat(info.latitude);
-						if (latError != null)
-							setError(latError);
-						break;
-
-					case LON_COL:
-						String longStr = ((String) aValue).trim();
-						double doubleLong = 0;
-						String dir2 = longStr.substring(longStr.length() - 1, longStr.length());
-						if (dir2.equalsIgnoreCase("E") || dir2.equalsIgnoreCase("W")) {
-							if (longStr.length() > 2) {
-								doubleLong = Double.parseDouble(longStr.substring(0, longStr.length() - 1));
-								doubleLong = Math.round(doubleLong * 100.0) / 100.0;
-								info.longitude = doubleLong + " " + dir2.toUpperCase();
-							}
-							else {
-								info.longitude = (String) aValue;
-							}
-						}
-						else {
-							info.longitude = (String) aValue;
-						}
-						String lonError = Coordinates.checkLon(info.longitude);
-						if (lonError != null)
-							setError(lonError);
-						break;
-					}
-				}
-
-				if (columnIndex != SPONSOR_COL || columnIndex != PHASE_COL)
-					checkForAllErrors();
-
-				fireTableDataChanged();
-			}
-		}
-
-		/**
-		 * Remove a set of settlements from the table.
-		 * 
-		 * @param rowIndexes
-		 *            an array of row indexes of the settlements to remove.
-		 */
-		private void removeSettlements(int[] rowIndexes) {
-			List<SettlementInfo> removedSettlements = new ArrayList<SettlementInfo>(rowIndexes.length);
-
-			for (int x = 0; x < rowIndexes.length; x++) {
-				if ((rowIndexes[x] > -1) && (rowIndexes[x] < getRowCount())) {
-					removedSettlements.add(settlementInfoList.get(rowIndexes[x]));
-				}
-			}
-
-			Iterator<SettlementInfo> i = removedSettlements.iterator();
-			while (i.hasNext()) {
-				SettlementInfo s = i.next();
-				settlementInfoList.remove(s);
-			}
-
-			fireTableDataChanged();
-		}
-
-		/**
-		 * Adds a new settlement to the table.
-		 * 
-		 * @param settlement
-		 *            the settlement configuration.
-		 */
-		private void addSettlement(InitialSettlement settlement) {
-			settlementInfoList.add(toSettlementInfo(settlement));
-			fireTableDataChanged();
-		}
-
-		/**
-		 * Check for all errors in the table.
-		 */
-		private void checkForAllErrors() {
-			clearError();
-			
-			Set<Coordinates> usedCoordinates = new HashSet<>();
-			Set<String> usedCrews = new HashSet<>();
-			for(SettlementInfo settlement : settlementInfoList) {
-
-				// Check that settlement name is valid.
-				if ((settlement.name == null) || (settlement.name.isEmpty())) {
-					setError(Msg.getString("SimulationConfigEditor.error.nameMissing")); //$NON-NLS-1$
-				}
-
-				// Check if population is valid.
-				if ((settlement.population == null) || (settlement.population.isEmpty())) {
-					setError(Msg.getString("SimulationConfigEditor.error.populationMissing")); //$NON-NLS-1$
-				} else {
-					try {
-						int popInt = Integer.parseInt(settlement.population);
-						if (popInt < 0) {
-							setError(Msg.getString("SimulationConfigEditor.error.populationTooFew")); //$NON-NLS-1$
-						}
-					} catch (NumberFormatException e) {
-						setError(Msg.getString("SimulationConfigEditor.error.populationInvalid")); //$NON-NLS-1$
-					}
-				}
-
-				// Check if number of robots is valid.
-				if ((settlement.numOfRobots == null) || (settlement.numOfRobots.isEmpty())) {
-					setError(Msg.getString("SimulationConfigEditor.error.numOfRobotsMissing")); //$NON-NLS-1$
-				} else {
-					try {
-						int num = Integer.parseInt(settlement.numOfRobots);
-						if (num < 0) {
-							setError(Msg.getString("SimulationConfigEditor.error.numOfRobotsTooFew")); //$NON-NLS-1$
-						}
-					} catch (NumberFormatException e) {
-						setError(Msg.getString("SimulationConfigEditor.error.numOfRobotsInvalid")); //$NON-NLS-1$
-					}
-				}
-				
-				String latError = Coordinates.checkLat(settlement.latitude);
-				if (latError != null)
-					setError(latError);
-				String lonError = Coordinates.checkLon(settlement.longitude);
-				if (lonError != null)
-					setError(lonError);
-				
-				// Only check duplicate if no errors
-				if ((latError == null) && (lonError == null)) {
-					Coordinates c = new Coordinates(settlement.latitude, settlement.longitude);
-					if (!usedCoordinates.add(c))
-						setError(Msg.getString("Coordinates.error.latitudeLongitudeRepeating")); //$NON-NLS-1$
-				}
-				
-				if (settlement.crew != null) {
-					if (usedCrews.contains(settlement.crew)) {
-						setError(Msg.getString("SimulationConfigEditor.error.duplicateCrew"));
-					}
-					else {
-						usedCrews.add(settlement.crew);
-					}
-				}
-			}
-		}
-	}
-	
 	/** default logger. */
 	private static final SimLogger logger = SimLogger.getLogger(SimulationConfigEditor.class.getName());
 
 	private static final int HORIZONTAL_SIZE = 1024;
-	
-	private static final int SETTLEMENT_COL = 0;
-	private static final int SPONSOR_COL = 1;
-	private static final int PHASE_COL = 2;
-	private static final int SETTLER_COL = 3;
-	private static final int CREW_COL = 4;
-	private static final int BOT_COL = 5;
-	private static final int LAT_COL = 6;
-	private static final int LON_COL = 7;
 
-	private static final int NUM_COL = 8;
 	
 	// Data members.
 	private boolean hasError, isCrewEditorOpen = true;
@@ -587,7 +155,7 @@ public class SimulationConfigEditor {
 	private Font DIALOG_14 = new Font("Dialog", Font.PLAIN, 14);
 	private Font DIALOG_16 = new Font("Dialog", Font.BOLD, 16);
 	
-	private SettlementTableModel settlementTableModel;
+	private InitialSettlementModel settlementTableModel;
 	private JTable settlementTable;
 	private JLabel errorLabel;
 	private JButton startButton;
@@ -712,18 +280,18 @@ public class SimulationConfigEditor {
 		f.add(settlementScrollPane, BorderLayout.CENTER);
 
 		// Create settlement table.
-		settlementTableModel = new SettlementTableModel();
+		settlementTableModel = new InitialSettlementModel(settlementConfig, raFactory);
 		
 		settlementTable = new JTable(settlementTableModel);
 		settlementTable.setRowSelectionAllowed(true);
-		settlementTable.getColumnModel().getColumn(SETTLEMENT_COL).setPreferredWidth(80);
-		settlementTable.getColumnModel().getColumn(SPONSOR_COL).setPreferredWidth(80);
-		settlementTable.getColumnModel().getColumn(PHASE_COL).setPreferredWidth(40);
-		settlementTable.getColumnModel().getColumn(SETTLER_COL).setPreferredWidth(30);
-		settlementTable.getColumnModel().getColumn(CREW_COL).setPreferredWidth(30);
-		settlementTable.getColumnModel().getColumn(BOT_COL).setPreferredWidth(30);
-		settlementTable.getColumnModel().getColumn(LAT_COL).setPreferredWidth(35);
-		settlementTable.getColumnModel().getColumn(LON_COL).setPreferredWidth(35);
+		settlementTable.getColumnModel().getColumn(InitialSettlementModel.SETTLEMENT_COL).setPreferredWidth(80);
+		settlementTable.getColumnModel().getColumn(InitialSettlementModel.SPONSOR_COL).setPreferredWidth(80);
+		settlementTable.getColumnModel().getColumn(InitialSettlementModel.PHASE_COL).setPreferredWidth(40);
+		settlementTable.getColumnModel().getColumn(InitialSettlementModel.SETTLER_COL).setPreferredWidth(30);
+		settlementTable.getColumnModel().getColumn(InitialSettlementModel.CREW_COL).setPreferredWidth(30);
+		settlementTable.getColumnModel().getColumn(InitialSettlementModel.BOT_COL).setPreferredWidth(30);
+		settlementTable.getColumnModel().getColumn(InitialSettlementModel.LAT_COL).setPreferredWidth(35);
+		settlementTable.getColumnModel().getColumn(InitialSettlementModel.LON_COL).setPreferredWidth(35);
 		settlementTable.setSelectionMode(ListSelectionModel.MULTIPLE_INTERVAL_SELECTION);
 		settlementTable.setBackground(java.awt.Color.WHITE);
 
@@ -732,20 +300,20 @@ public class SimulationConfigEditor {
 		settlementScrollPane.setViewportView(settlementTable);
 
 		// Create combo box for editing sponsor column in settlement table.
-		TableColumn sponsorColumn = settlementTable.getColumnModel().getColumn(SPONSOR_COL);
+		TableColumn sponsorColumn = settlementTable.getColumnModel().getColumn(InitialSettlementModel.SPONSOR_COL);
 		WebComboBox sponsorCB = new WebComboBox();
 		sponsorCB.setModel(new UserConfigurableComboModel(raFactory, false));
 		sponsorColumn.setCellEditor(new DefaultCellEditor(sponsorCB));
 		
 		// Create combo box for editing crew column in settlement table.
 		// Use a custom model to inherit new Crews
-		TableColumn crewColumn = settlementTable.getColumnModel().getColumn(CREW_COL);
+		TableColumn crewColumn = settlementTable.getColumnModel().getColumn(InitialSettlementModel.CREW_COL);
 		JComboBoxMW<String> crewCB = new JComboBoxMW<String>();
 		crewCB.setModel(new UserConfigurableComboModel(crewConfig, true));
 		crewColumn.setCellEditor(new DefaultCellEditor(crewCB));
 		
 		// Create combo box for editing template column in settlement table.
-		TableColumn templateColumn = settlementTable.getColumnModel().getColumn(PHASE_COL);
+		TableColumn templateColumn = settlementTable.getColumnModel().getColumn(InitialSettlementModel.PHASE_COL);
 		JComboBoxMW<String> templateCB = new JComboBoxMW<String>();
 		for (SettlementTemplate st : settlementConfig.getSettlementTemplates()) {
 			templateCB.addItem(st.getTemplateName());
@@ -757,7 +325,7 @@ public class SimulationConfigEditor {
 		DefaultTableCellRenderer defaultTableCellRenderer = new DefaultTableCellRenderer();
 		defaultTableCellRenderer.setHorizontalAlignment(SwingConstants.LEFT);
 		TableColumn column = null;
-		for (int ii = 0; ii < NUM_COL; ii++) {
+		for (int ii = 0; ii < settlementTableModel.getColumnCount(); ii++) {
 			column = settlementTable.getColumnModel().getColumn(ii);
 			column.setCellRenderer(defaultTableCellRenderer);
 		}
@@ -798,6 +366,15 @@ public class SimulationConfigEditor {
 		errorLabel = new JLabel("", JLabel.CENTER); //$NON-NLS-1$
 		errorLabel.setForeground(Color.RED);
 		bottomPanel.add(errorLabel, BorderLayout.NORTH);
+		
+		// Monitor table models for errors
+		settlementTableModel.addTableModelListener(new TableModelListener() {
+			
+			@Override
+			public void tableChanged(TableModelEvent e) {
+				checkModelErrors();
+			}
+		});
 		
 		// Create the config control
 		configControl = new UserConfigurableControl<Scenario>(f, "Scenario", scenarioConfig) {
@@ -911,13 +488,34 @@ public class SimulationConfigEditor {
 		rootPane.setDefaultButton(startButton);
 	}
 	
-
+	/**
+	 * Check the table models for errors
+	 */
+	private void checkModelErrors() {
+		String settlementErrors = settlementTableModel.getErrorMessage();
+		if (settlementErrors != null) {
+			if (!hasError) {
+				hasError = true;
+				errorLabel.setText(settlementErrors);
+				startButton.setEnabled(false);
+				configControl.allowSaving(false);
+			}
+		}
+		else {
+			clearError();
+		}
+	}
+	
 	/**
 	 * Adds a new settlement with default values.
 	 */
 	private void addDefaultNewSettlement() {
-		InitialSettlement settlement = determineNewDefaultSettlement();
-		settlementTableModel.addSettlement(settlement);
+		// Get any known Reporting Authority
+		String sponsor = raFactory.getItemNames().iterator().next();
+		String template = determineNewSettlementTemplate();
+		Coordinates location = determineNewSettlementLocation();
+
+		settlementTableModel.addPartialSettlement(sponsor, template, location);
 	}
 
 	/**
@@ -987,21 +585,6 @@ public class SimulationConfigEditor {
 	}
 	
 	/**
-	 * Sets an edit check error.
-	 * 
-	 * @param errorString
-	 *            the error description.
-	 */
-	private void setError(String errorString) {
-		if (!hasError) {
-			hasError = true;
-			errorLabel.setText(errorString);
-			startButton.setEnabled(false);
-			configControl.allowSaving(false);
-		}
-	}
-
-	/**
 	 * Clears all edit-check errors.
 	 */
 	private void clearError() {
@@ -1011,25 +594,6 @@ public class SimulationConfigEditor {
 			startButton.setEnabled(true);
 			configControl.allowSaving(true);
 		}
-	}
-
-	/**
-	 * Configures a new default settlement.
-	 * 
-	 * @return settlement configuration.
-	 */
-	private InitialSettlement determineNewDefaultSettlement() {
-		
-		// Get any known Reporting Authority
-		String sponsor = raFactory.getItemNames().iterator().next();
-		String template = determineNewSettlementTemplate();
-		Coordinates location = determineNewSettlementLocation();
-		
-		return new InitialSettlement(tailorSettlementNameBySponsor(sponsor,
-										settlementTableModel.getRowCount()),
-				sponsor, template, 
-				determineNewSettlementPopulation(template),
-				determineNewSettlementNumOfRobots(template), location, null);
 	}
 
 	/**
@@ -1050,83 +614,6 @@ public class SimulationConfigEditor {
 		return result;
 	}
 
-	/**
-	 * Determines the new settlement population.
-	 * 
-	 * @param templateName
-	 *            the settlement template name.
-	 * @return the new population number.
-	 */
-	private int determineNewSettlementPopulation(String templateName) {
-
-		int result = 0; 
-
-		if (templateName != null) {
-			Iterator<SettlementTemplate> i = settlementConfig.getSettlementTemplates().iterator();
-			while (i.hasNext()) {
-				SettlementTemplate template = i.next();
-				if (template.getTemplateName().equals(templateName)) {
-					result = template.getDefaultPopulation();
-				}
-			}
-		}
-
-		return result;
-	}
-
-	/**
-	 * Determines the new settlement number of robots.
-	 * 
-	 * @param templateName
-	 *            the settlement template name.
-	 * @return number of robots.
-	 */
-	private int determineNewSettlementNumOfRobots(String templateName) {
-
-		int result = 0;
-
-		if (templateName != null) {
-			Iterator<SettlementTemplate> i = settlementConfig.getSettlementTemplates().iterator();
-			while (i.hasNext()) {
-				SettlementTemplate template = i.next();
-				if (template.getTemplateName().equals(templateName)) {
-					result = template.getDefaultNumOfRobots();
-
-				}
-			}
-		}
-
-		return result;
-	}
-
-	/**
-	 * Returns a random settlement name tailored by the sponsor
-	 * 
-	 * @param sponsor
-	 * @return
-	 */
-	private String tailorSettlementNameBySponsor(String sponsor, int index) {
-		ReportingAuthority ra = raFactory.getItem(sponsor);
-
-		List<String> usedNames = new ArrayList<>();
-		
-		// Add configuration settlements from table data.
-		for (int x = 0; x < settlementTableModel.getRowCount(); x++) {
-			String name = (String) settlementTableModel.getValueAt(x, SETTLEMENT_COL);
-			usedNames.add(name);
-
-		}
-
-		// Gets a list of settlement names that are tailored to this country
-		List<String> candidateNames = new ArrayList<String>(ra.getSettlementNames());
-		candidateNames.removeAll(usedNames);
-
-		if (candidateNames.isEmpty())
-			return "Settlement #" + index;
-		else
-			return candidateNames.get(RandomUtil.getRandomInt(candidateNames.size()-1));
-	}
-	
 	/** 
 	 * Get a new Location ofr a Settlement 
 	 * @return
